libs/common/pyproject.toml
[project]
name = "hive-common"
dynamic = ["version"]
authors = [{ name = "Gary Benson", email = "gary@gbenson.net" }]
description = "Common code for Hive libraries and services"
readme = "README.md"
license = "MIT"
dependencies = [
    "PyYAML",
]

[project.urls]
Homepage = "https://github.com/gbenson/hive/tree/main/libs/common"
Source = "https://github.com/gbenson/hive"

[project.optional-dependencies]
brotli = [
    "httpx[brotli]",
]
httpx = [
    "hishel",
    "httpx[http2]",
]
zstd = [
    "httpx[zstd]",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {attr = "hive.common.__version__.__version__"}

[tool.mypy]
warn_unused_configs = true
strict = true

[tool.pytest.ini_options]
addopts = "--cov=hive.common"
log_level="debug"

[tool.coverage.run]
omit = [
    "*/venv/*",
    "*/.venv/*",
    "*/hive/*/__main__.py",
]

[tool.coverage.report]
skip_covered = true
